fp16 as GPU not support BF16

+++ dpo_train.py
@@ -39,7 +39,7 @@ test_data = dataset_split["test"]
 # setup DPO training
 training_args = DPOConfig(
     output_dir="qomhra-8B-awq-dpo", 
-    bf16=True, # unsloth
+    fp16=True, # bf16 not supported by v100
     per_device_train_batch_size=1, # OOM
     num_train_epochs=3,
     learning_rate=5e-7,
